fix: remove all files/folders from last run
Consider the behavior of the bookkeper's DISCARD_FULL mode when
`RUN = 1-3 1` (or more complex, i.e., more repetitions):

The current implementation by @amimre for DISCARD_FULL would remove:
- ONLY the "primary" history directory, i.e., the one created from the very last segment of calc. I.e., the folders archived from workhistory would remain.
- No Tensors and Deltas, as the logic was flawed: it intended to remove ONLY the ones with the index corresponding to the last refcalc segment in the last run (i.e., the last '1' above), but would remove none at all because now we do archive automatically (so tensor_num would always be among the one counted in max_job_per_tensor!).

The old implementation of bookkeeper --no-cont  by @fkraushofer would remove:
- All the directories created during the last calc run, i.e., BOTH the "primary" output as well as all the intermediate ones stored in workhistory. This is because none of the folders were stored in history yet.
- ONLY the Tensors/Deltas with an index corresponding to the LAST refcalc segment of the last calc run. In the example above they would be those generated in the last '1'.

I believe that the correct behavior is to remove:
- All the directories created during the last calc run, i.e., BOTH the "primary" output as well as all the intermediate ones that were originally stored in workhistory. (We now archive all of them automatically.)
- ALL the Tensors/Deltas that only involve the last calc run, i.e., both those generated in the first '1-3' as well as those of the last '1'.

This commit implements the last behavior. I would like to hear the opinion of @fkraushofer and @amimre on this. Do you agree that this is the correct behavior? Also, please cross-check my current logics uses to decide whether a Tensor (and its Deltas) should be removed: I consider a Tensor as "new" if there is at most one "job number" corresponding to it. Only (and all) the tensor numbers that were archived in the last run  are considered, i.e., both the primary one as well as those from workhistory. Can you think of a setting for RUN that would create more than one "job number" and would still require removal of those Tensors/Deltas?