fix: restore parameters.write path kwarg
As discussed in point 2 of https://github.com/viperleed/viperleed/pull/198#issuecomment-2474020727. The use of rpars.workdir as the path where PARAMETERS are written is problematic, as for domains calculations we pass along dp.rp, which is (correctly) set to the main work directory rather than the work directory in which the specific domain is handled.