fix,tests: non-reproducible test failure
The reason for the test failure was the absence of a log file in the root directory. This would cause the bookkeeper to fall back to the '-moved_<now>' suffix, which could have differed by one second between L715 (storage of the soon-to-be history folder) and L726 (running the bookkeper again), causing the creation of a folder with a different name. The same test was also flawed in the following --discard-full run, which would be delayed even further.