fix: never fall back to _ori files
Looking back at my description in the header of #198, I realized that the following makes no sense:

> [...]
> #### Mode `--clear`/`-c`
This is the mode in which bookkeeper runs before `run_calc`.
> 1. Do the same as mode `--archive`, **but --- when making a history folder if none exists --- copy the _ori-suffixed files as inputs removing the _ori suffix. This is just in case bookkeeper failed to run at the end of the previous calc run.
> [...]

The problem with this portion of the behavior is that there is no sensible condition in which this would ever happen. In fact, creation of `_ori`-suffixed files only happens in ARCHIVE mode, **after** the current directory has been archived to history. However, modes CLEAR and DISCARD do correctly return early if there is already an archived directory. This means that _copy_input_files_from_original_inputs_or_cwd (which is part of making a new history subfolder) never executes in these modes as soon as ARCHIVE created a folder. This is irrespective of the contents of the folder, even if ARCHIVE failed to move some files.

A discussion with @amimre a few days ago confirmed my suspicion. It would be great if @fkraushofer could think about this too and provide his feedback.

This commit does remove the relevant portions of the code, and adds explicit log messages in case ARCHIVE fails to handle any of the input_file->_ori + OUT->new_input_file.