fix: don't catch-all
organize_workdir already handles sensible exceptions internally. If some exception is raised, it is because of a bug in the code, which we should not mask. In fact, it is likely going to affect collection of files from all the directories. It then makes more sense to fail already at the first occurrence.