fix: update readParams upon edits
This ensures that checks like "param in readParams" always return the correct result, even when a parameter was commented out and the PARAMETERS file is not read back in. Also, it ensures that .readParams always contains up-to-date information concerning the most recent PARAMETERS file.

In order for this to work correctly, the internal behavior of ParametersEditor is changed to store tuples of (parameter_name, original_assignment) rather than parameter_name alone. This has the added benefit of (soon) supporting edits of parameters that may be given multiple times and for which all the assignments are considered (rather than only using the last one). This is the case for DOMAIN, whose edits need to be supported to fix #301 in #305.