feat: finish implementing multi-value support
For now, the new implementation is not tested yet, as ModifiedParameterValue does not yet allow edits of any multi-value parameter. Will soon be the case.

The major change in this commit concerns the return value of parameters.modify, which now returns a ModifiedParameterValue rather than its .fmt_value directly. This is critical to support editing DOMAIN.