fix: z_periodic=True by default
The root cause of #306 has to do with the different default values for the `z_periodic` kwarg in `get_minimal_c_vector(..., z_periodic=True)` vs `ensure_minimal_c_vector(..., z_periodic=False)`.

The former method was used in `findBulkSymmetry` before 5c3df0eace0de07b99985b84cad8ce4091c660c1, and was refactored to use the latter in that commit.

Given the fact that the default value for `z_periodic` is `True` in all the other functions, it feels weird to have `ensure_minimal_c_vector` make an exception.

Notice that the use of `z_periodic=True` in `findBulkSymmetry` is correct, as the bulk slab that is passed along at that point has already been reduced along c after detecting the bulk (in `initialization`).