fix: modify only once multi-valued
This is perhaps not ideal if we ever plan to support modification of stuff like SITE_DEF, where, yes, multiple assignments are considered, but only the last assignment for each specific flag value is taken. However, that case should probably be handled at _infer_original_assignment, which should return the unused values in the "to be commented" part.