fix: consistent .path attributes
We now have a few classes that handle folders or files. Make sure that the .path attribute is always a Path pointing to the object handled.

This was not the case for BookkeeperMetaFile and ManifestFile, where .path would point to the containing folder.

Now their attributes are made clearer:
* BookkeeperMetaFile has a .folder attribute for the path to the folder it is contained in
* ManifestFile has, instead a .root attribute (since it refers to both the folder that contains it as well as the top-level contents). Its former .paths attribute has been renamed to .sections (in line with the .iter_sections() method).
* Both now have a convenience .file alias for .path.

Here all the classes that handle files/folders up to now (and have a .path attribute):
* folder: WorkhistoryHandler, HistoryExplorer, (Incomplete)HistoryFolder, RootExplorer
* file: HistoryInfoFile, ManifestFile, BookkeeperMetaFile
* file or folder: TensErLEEDSource