feat: warn when running in an old tree
We do have some form of backward compatibility (e.g., handling old calc log names, handling of _OUT files), but many things have changed in, e.g., the storage of original_inputs, the mechanics of how bookkeeper treated files, and the "default mode" for bookkeeper. Running in an tree that was created by an old version of calc (and bookkeeper) may not always yield the expected results.