feat: StreamArgument
This is an improved version of argparse's FileType to be used to turn an argument into an open stream, but only when entering a context. Instead, FileType opens file paths at creation rather than at use. This was the cause of #285, which was fixed in #286. The solution added here is of more general use, as StreamArgument can be used in other places.