feat: support --discard-full on old tree
In order to correctly remove runs via --discard-full in a tree created by an older bookkeeper version, we need to have the correct metadata files in place. Since the timestamp is a unique identifier for runs that were created at the same time, we can add the metadata files correctly as long as: folders have been created by calc (no "moved-") and there is only one "main" folder with one timestamp.