feat: Workhistory takes the full root explorer
This has the advantage of decoupling the WorkhistoryHandler code from the bookkeeper instance as much as possible. The only exception is the .timestamp attribute. This change is needed in order to support multi-domain execution, where RootExplorer, rather than Bookkeeper, will contain information on domain subfolders. This way, WorkhistoryHandler will always access the history of the specific domain subfolder (or the main one).