feat: DomainFinder
Basic implementation of a DomainFinder class. Replaces functionality that used to be in RootExplorer to better confine code. Also, RootExplorer does not use any longer the DomainRootExplorer class, which is, for now, kept in place. The reason is not to overdo the collection of information from domain subfolders, as the current implementation only ever needs to check whether there are any domain subfolders.

The DomainFinder class will be expanded shortly to also find true domains, not only potential ones.