fix: no log file in domain subfolders
The previous implementation was wrong, as domain subfolders do not contain a log file, and must thus rely on the log file of the main root directory. This includes timestamp (which goes to the history subfolders as well as history.info) and run segments (for history.info).

This commit fixes the problem by creating a dedicated Bookkeeper subclass that has a DomainRootExplorer as its _root private attribute rather than a "normal" RootExplorer.