feat: preemptively check DISCARD_FULL
For a multi-domain calculation, check beforehand whether it is possible to DISCARD_FULL an entry from both the root as well as all domains. The checks that should pass are those that concern the "removability" of the entry. There is still one case that can't be handled beforehand: an OSError when trying to remove files/folders. This can't be reasonably checked in advance.