fix: flaw in DISCARD_FULL
The previous implementation would NOT remove the most recent Tensors/Deltas, but **the ones of the previous run**. In fact, by the time remove_tensors_and_deltas was called, the most-recent history run would be the **previous** one, as the one to be deleted was deleted already. Also, the missing update of the history contents after removal (fixed in the previous commit) would likely mask the problem.