fix: clarify that -y overrides user confirmation
As pointed out by @fkraushofer in https://github.com/viperleed/viperleed/pull/325#pullrequestreview-2820898038, the documentation for the '-y' CLI argument was incorrect.

Added some tests to clarify that user confirmation is fully skipped in case '-y' is given as a CLI argument. And to clarify that whether the shell is interactive or not is irrelevant.