feat: fallback for missing unlabled inputs
If, after ARCHIVE, there are no unlabeled files corresponding to any of the _ori ones, duplicate the latter. This is a fallback, and should never happen, as discussed in https://github.com/viperleed/viperleed/issues/353#issuecomment-2886613673

The reason for doing this only in ARCHIVE mode is that ARCHIVE is the only mode that produces _ori files in the first place.