fix: don't discard if physically acceptable
When editing the exposure time of a camera, assigning an out-of-range value in the spinbox would cause the user value to be discarded completely rather than coerced to the expected limits.

This is because calling setMinimum on the widget makes out-of-range values be considered invalid. (Discarding input is the default behavior for invalid values.) Instead, to better fit the user expectations, use a physical limit for the minimum value that is considered unacceptable. This allows users (who do not knw that, e.g., IS cameras can do no more than 20us) to set 0 in the box and get the minimum.

Negative values would still be considered invalid and discarded completely.