refactor: rename Atom.known_deltas
After discussion with @fkraushofer, we realized that the name was not really fitting. Indeed, those are the delta files used in the current search iteration, while the known_deltas name made it sound as if there may be more. The reason why it is only the "current" ones is in SurfaceSlab.restoreOriState, which is called in run_sections after a superpos.