fix: tread subdomains the same as single
Before this commit we would add the Deltas folder to manifest differently for a single-domain calculation (always added on successful completion of all calculations) than for a subdomain of a multi-domain calculation (added only if there was anything to calculate freshly, but before the actual calculation). Now, the folder is added in both cases irrespective of whether the calculation is successful, as long as any calculation is done. The only drawback of this is that, should calculations fail and not create a Deltas folder at all, an error message will be printed about failure to copy the (missing) Deltas folder back to home.