fix: do not re-write same parameter value
calling parameters.modify with an identical value would duplicate the edited value. This was not a primary issue for "normal" parameters as the latest value overrides the previous ones (though it may have confused users in long PARAMETERS files). However, it is a major problem for DOMAIN, as we complain if there are multiple lines referencing a domain with the same "flag". This bug would render PARAMETERS files on which DOMAIN was given essentially not reusable (e.g., RUN=0 then RUN=4 would fail at reading PARAMETERS).