fix: ensure domain Tensors have IVBEAMS
When a refcalc is executed in a multi-domain tree, the IVBEAMS file from the main work directory needs to be injected in the Tensor.zip files. Otherwise RUN=0 after RUN=1 would still identify domains as requiring a refcalc, even if nothing changed.

Also, clarify where IVBEAMS comes from in the documentation.

Co-Authored-By: Florian Kraushofer <66359721+fkraushofer@users.noreply.github.com>