fix: write inherited PARAMETERS in domains
Ensure that any PARAMETERS that domains inherit from the main ones are explicitly written to the relevant PARAMETERS files. This concerns domains for which a refcalc needs to be executed and those PARAMETERS that were either not given by users or whose value is inconsistent with the one of the main folder.

Also:
- refactor into dedicated functions for improving readability
- ensure that IVBEAMS and the PARAMETERS modification only affects those domains that require a new refcalc. In the previous version, all domains would be affected if any of them required to be recalculated.