fix: domains always inherit TL_VERSION
This must happen (i) for all domains (irrespective of whether a refcalc needs to be done), and (ii) before calling initialization for the domains, as, if not given explicitly, it would be otherwise auto-detected to the most recent one.

Do a hard override of user-given values as all domains need to be processed with the same version (and the same of the main) in order to be successfully "combined" in a search.