fix: prevent duplicates in Tensors.zip
The old implementation would cause Tensors.zip (and Deltas.zip, besides the addition of new deltas) to grow in size proportionally to the number of times they were used in a calc execution.

This is because ZipFile adds archive items with conflicting names to an existing archive (only issuing a UserWarning about it) without any checking for actual duplication.

This commit skips items with the same name, under the assumption that they are the same file. As discussed with @fkraushofer, this should always be the case (aside from twisted corner cases of re-usage of the same work folder for two different calculations).