feat: safely detect crashes on PyQt5 imports
After a bit of analysis, it turns out that Python may crash already at the time PyQt5 or any of its subpackages are imported (different packages depend on different sets of pre-compiled runtime libraries). This typically signals cases of ABI mismatch between the Qt that PyQt5 loads and other Qts that may be present on the system. Anaconda is a very common cause of this, as it injects itself into LD_LIBRARY_PATH.

Crashes may cause Python ImportErrors, or even qFatal in the Qt core code. The latter is the cause of the bare "Aborted (core dumped)" message.

This is likely the cause of the problems seen by @Garlic-King in #434.

This commit makes PyQt5 imports safe by trying them first into a sandboxed subprocess. The previous code would only check for runtime crashes resulting from creating a QApplication.

Attempting to start the GUI will display an understandable message rather than a bare "Aborted (core dumped)".