fix: don't search for libs if none is missing
The previous implementation, especially the one in _list_install_for_libs_linux, would be problematic if called with no missing dependencies: apt-cache search '' would list everything. As this problem may re-appear for other platforms that we may implement in the future, it is best to shortcut in the dispatcher if no dependency is missing.