fix: refcalc and deltas with Intel on Windows
On Windows, subprocess (which uses spawn) does not create a proper STDIN stream when used with multiprocessing. This means one cannot pipe to subprocesses the contents of input files. Instead, proper file handles need to be passed. This commits ensures that the input for refcalc and delta amplitude are successfully written to file, and passes those file handles. This should have no impact when working on Unix (other than a very minor I/O overhead for deltas).