fix: make search work on Windows
Tested only with ifort/mpiifort, as getting gfortran with lapack/blas & mpi on native windows is a pain.

This commits re-structures the search_job module to introduce:
- A SearchWorkerABC abstract class that incorporates the platform-independent behavior.
- Providing concrete implementations for Windows and Unix.

The Unix implementation is essentially identical to the previous code. On Windows, some shenanigans using the Windows API DLLs are needed to ensure proper killing of the search process. See comments at _create_winapi_handles for more details.

Notes: the current implementation for Windows does not work correctly for older versions of the search (at least with the intel compilers) where rf.info is piped to the fortran code. I tried an implementation similar to the one in 9b4dc88f5b1ce4f52951a551d0912c491759bfcb, but that still seems not be enough. We fail early now and point the users to the solution. The minimum version suggested is the first one that uses Fortran's random-number generation to prevent the hassle of rewriting and manually compiling random_.c (see #151).